README
A Project on School website using HTML, CSS, JavaScript and Bootstrap. It is a website where people can check details about the school like the facilities, events and activities etc. all the information you need to know is provided here. It is a responsive website, where one can drop in some doubts and queries. 
Animation is used to make the website aesthetically pleasing.